+++ examples/sudoku/timer.h
@@ -18,13 +18,13 @@ class Timer {
   void reset() { start_ = 0; end_ = 0; rounds_ = 0; }
 
   double duration() const { return (end_ - start_) / (double) CLOCKS_PER_SEC; }
-  size_t rounds() const { return rounds_; }
+  int rounds() const { return rounds_; }
   double avg_duration() const { return duration() / rounds_; }
 
  private:
   std::clock_t start_;
   std::clock_t end_ = 0;
-  size_t rounds_ = 0;
+  int rounds_ = 0;
 };
 
 #endif  // EXAMPLES_SUDOKU_TIMER_H_
